old dialog deleted, material imports put in their own module
