+++ package.json
@@ -32,6 +32,7 @@
     "command-line-args": "^5.1.1",
     "core-js": "^2.5.4",
     "express": "^4.16.4",
+    "hammerjs": "^2.0.8",
     "moment": "^2.24.0",
     "mongoose": "^5.6.4",
     "ng2-charts": "^2.2.2",
