+++ src/app/app.component.html
@@ -14,9 +14,9 @@
 		<mat-sidenav class="side-navigation" #sideNavigation opened mode="side">
 			<div fxLayout="column" fxLayoutAlign="start space-between" fxFlex>
 				<mat-nav-list fxLayout="column" fxLayoutAlign="start stretch">
-          <!-- <a mat-list-item [routerLink]="['/food']">Food</a>
-          <a mat-list-item [routerLink]="['/body-mind']">Body & Mind</a>
-          <a mat-list-item [routerLink]="['/money-manager']">Money Manager</a> -->
+					<!-- <a mat-list-item [routerLink]="['/food']">Food</a>
+					<a mat-list-item [routerLink]="['/body-mind']">Body & Mind</a> -->
+					<a mat-list-item [routerLink]="['/money-manager']">Money Manager</a>
 					<a mat-list-item [routerLink]="['/workflow-manager']">Workflow Manager</a>
 					<!-- <a mat-list-item [routerLink]="['/long-term-goals']">Long Term Goals</a> -->
 				</mat-nav-list>
