+++ src/app/app.component.scss
@@ -7,6 +7,10 @@
   height: 65px;
 }
 
+.manager-toolbar {
+  height: 50px;
+}
+
 .side-navigation-container {
   height: calc(100% - 65px);
 }
