+++ src/app/common/material.module.ts
@@ -1,0 +1,101 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+
+import { A11yModule } from '@angular/cdk/a11y';
+import {
+	MatAutocompleteModule,
+	MatBadgeModule,
+	MatBottomSheetModule,
+	MatButtonModule,
+	MatButtonToggleModule,
+	MatCardModule,
+	MatCheckboxModule,
+	MatChipsModule,
+	MatDatepickerModule,
+	MatDialogModule,
+	MatDividerModule,
+	MatExpansionModule,
+	MatGridListModule,
+	MatIconModule,
+	MatInputModule,
+	MatListModule,
+	MatMenuModule,
+	MatNativeDateModule,
+	MatPaginatorModule,
+	MatProgressBarModule,
+	MatProgressSpinnerModule,
+	MatRadioModule,
+	MatRippleModule,
+	MatSelectModule,
+	MatSidenavModule,
+	MatSliderModule,
+	MatSlideToggleModule,
+	MatSnackBarModule,
+	MatSortModule,
+	MatStepperModule,
+	MatTableModule,
+	MatTabsModule,
+	MatToolbarModule,
+	MatTooltipModule,
+	MatTreeModule
+} from '@angular/material';
+
+const materialModules = [
+	A11yModule, // ARIA accessibility
+	// CdkTableModule,
+	// CdkTreeModule,
+	// DragDropModule,
+	MatAutocompleteModule,
+	// MatBadgeModule,
+	// MatBottomSheetModule,
+	MatButtonModule,
+	// MatButtonToggleModule,
+	MatCardModule,
+	MatCheckboxModule,
+	MatChipsModule,
+	// MatStepperModule,
+	MatDatepickerModule,
+	MatDialogModule,
+	MatDividerModule,
+	MatExpansionModule,
+	// MatGridListModule,
+	MatIconModule,
+	MatInputModule,
+	MatListModule,
+	// MatMenuModule,
+	// MatMomentDateModule,
+	MatNativeDateModule,
+	// MatPaginatorModule,
+	MatProgressBarModule,
+	// MatProgressSpinnerModule,
+	MatRadioModule,
+	// MatRippleModule,
+	MatSelectModule,
+	MatSidenavModule,
+	// MatSliderModule,
+	// MatSlideToggleModule,
+	// MatSnackBarModule,
+	MatSortModule,
+	MatTableModule,
+	MatTabsModule,
+	MatToolbarModule,
+	MatTooltipModule,
+	MatTreeModule
+	// ScrollingModule,
+];
+
+@NgModule({
+	declarations: [],
+	entryComponents: [],
+	exports: [
+		...materialModules,
+	],
+	imports: [
+		CommonModule,
+		...materialModules
+	],
+	providers: [
+	]
+})
+export class MaterialModule {
+}
